flask-pgsql/app.py
from flask import Flask, request, jsonify
import psycopg2

app = Flask(__name__)

def get_db_connection():
    conn = psycopg2.connect(
        host='db', # The service name of PostgreSQL in docker compose
        database='testdb',
        user='postgres',
        password='password'
    )
    return conn
@app.route('/')
def hello():
    return 'Hello, Docker Compose!'

@app.route('/users', methods=['GET'])
def get_users():
    conn = get_db_connection()
    cursor = conn.cursor()
    cursor.execute('SELECT * FROM users;')
    users = cursor.fetchall()
    cursor.close()
    conn.close()
    return jsonify(users)

if __name__ == '__main__':
    app.run(host='0.0.0.0', port=5000)
